# orders-api/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "IntegraHub Orders API",
    "version": "1.0.0",
    "description": "Documentación completa para la gestión de pedidos e integración asíncrona."
  },
  "paths": {
...
    "/orders": {
      "get": {
        "summary": "Listar todos los pedidos",
        "responses": {
          "200": { "description": "Lista de pedidos recuperada" }
        }
      },
      "post": {
        "summary": "Crear un nuevo pedido (Flujo A)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": { "type": "string", "example": "uuid-v4-correlation-id" },
                  "customer_name": { "type": "string", "example": "Richie Quintana" },
                  "total_amount": { "type": "number", "example": 50.00 }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Pedido recibido y publicado en RabbitMQ" },
          "401": { "description": "No autorizado - Falta Token JWT" },
          "409": { "description": "Conflicto - ID de pedido duplicado" }
        }
      }
    },
...
    "/analytics": {
      "get": {
        "summary": "Obtener analítica de ventas (Flujo D)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Reporte de analítica generado" },
          "401": { "description": "No autorizado - Falta Token JWT" },
          "503": { "description": "Base de datos no disponible (Circuit Breaker abierto)" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}